fix(ci): preserve image bake coverage
